Cleaned up comments; moved utilization
