+++ src/lib/markdown/vulnerability.ts
@@ -1,6 +1,3 @@
-// Vulnerability → markdown renderer. Structurally typed so any Prisma payload
-// with these fields is compatible.
-
 import { deviceGroupMatchingsSummary } from "./device-group";
 import { type CanonicalRef, shortId, truncate } from "./shared";
